[Refactor] Lecture Converter 변경
